+++ src/parsing/expansion/test/tries/word.c
